[LV][FEAT] added packages for pdf graph generation
