+++ package.json
@@ -24,6 +24,8 @@
   "homepage": "https://github.com/lucaunsw/H13-redstone-goats#readme",
   "dependencies": {
     "@vercel/node": "^2.3.0",
+    "canvas": "^3.1.0",
+    "chart.js": "^4.4.8",
     "cors": "^2.8.5",
     "csv-stringify": "^6.5.2",
     "dotenv": "^16.4.7",
